+++ src/Console/Command/Generator/PropertyComment/ModeA/Action/FindNamespaceDirectoryPath.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Action;
 
 use Ghjayce\Shipshape\Action\Action;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Entity\Context;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Entity\ModeAContext;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\PropertyCommentService;
 use Ghjayce\Shipshape\Entity\Context\ClientContext;
 use Ghjayce\Shipshape\Entity\Context\ShipshapeContext;
 
@@ -13,7 +14,7 @@ class FindNamespaceDirectoryPath extends Action
 {
 
     /**
-     * @param Context $context
+     * @param ModeAContext $context
      * @param ShipshapeContext $shipshapeContext
      * @return mixed
      */
@@ -21,7 +22,7 @@ class FindNamespaceDirectoryPath extends Action
     {
         $target = $context->getParam()->getTarget();
         if ($target[0] === '\\') {
-            $target = self::findNamespaceDirPath($target, $context->getParam()->getClassLoader());
+            $target = PropertyCommentService::findNamespaceDirPath($target, $context->getParam()->getClassLoader());
             if (!$target) {
                 throw new \RuntimeException("Class namespace '{$target}' not found");
             }