+++ src/Console/Command/Generator/PropertyComment/ModeA/Action/ReportScore.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Action;
 
 use Ghjayce\Shipshape\Action\Action;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Entity\Context;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\Original\PropertyCommentOriginal;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeA\Entity\ModeAContext;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\PropertyCommentService;
 use Ghjayce\Shipshape\Entity\Context\ClientContext;
 use Ghjayce\Shipshape\Entity\Context\ShipshapeContext;
 
@@ -14,13 +14,13 @@ class ReportScore extends Action
 {
 
     /**
-     * @param Context $context
+     * @param ModeAContext $context
      * @param ShipshapeContext $shipshapeContext
      * @return mixed
      */
     public function handle(ClientContext $context, ShipshapeContext $shipshapeContext): mixed
     {
-        PropertyCommentOriginal::reportScoreBoard($context->getScoreBoard());
+        PropertyCommentService::reportScoreBoard($context->getScoreBoard());
         return null;
     }
 }