+++ src/Console/Command/Generator/PropertyComment/ModeB/Action/EachClassWriteDocComment.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Action;
 
 use Ghjayce\Shipshape\Action\Action;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Entity\Context;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Entity\ModeBContext;
 use Ghjayce\Shipshape\Entity\Context\ClientContext;
 use Ghjayce\Shipshape\Entity\Context\ShipshapeContext;
 
@@ -13,12 +13,17 @@ class EachClassWriteDocComment extends Action
 {
 
     /**
-     * @param Context $context
+     * @param ModeBContext $context
      * @param ShipshapeContext $shipshapeContext
      * @return mixed
+     * @throws \JsonException
      */
     public function handle(ClientContext $context, ShipshapeContext $shipshapeContext): mixed
     {
-        // TODO: Implement handle() method.
+        $codeScoreBoard = $context->getService()->eachClassesWriteDocComment(
+            $context->getClassesWithNamespace(),
+            $context->getParam()->getIgnoreClassesWithNamespace()
+        );
+        return $context->setScoreBoard($codeScoreBoard);
     }
 }