+++ src/Console/Command/Generator/PropertyComment/ModeB/Action/ParamPreprocess.php
@@ -5,20 +5,25 @@ declare(strict_types=1);
 namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Action;
 
 use Ghjayce\Shipshape\Action\Action;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Entity\Context;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Entity\ModeBContext;
 use Ghjayce\Shipshape\Entity\Context\ClientContext;
 use Ghjayce\Shipshape\Entity\Context\ShipshapeContext;
 
-class FindNamespaceDirectoryPath extends Action
+class ParamPreprocess extends Action
 {
 
     /**
-     * @param Context $context
+     * @param ModeBContext $context
      * @param ShipshapeContext $shipshapeContext
      * @return mixed
      */
     public function handle(ClientContext $context, ShipshapeContext $shipshapeContext): mixed
     {
-        // TODO: Implement handle() method.
+        $target = $context->getParam()->getTarget();
+        $afterTarget = $context->getService()->targetPreprocess($target);
+        if (!$afterTarget) {
+            throw new \RuntimeException("'{$target}' is invalid");
+        }
+        return $context->setTarget($afterTarget);
     }
 }