+++ src/Console/Command/Generator/PropertyComment/ModeB/Contract/ServiceInterface.php
@@ -6,5 +6,10 @@ namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Cont
 
 interface ServiceInterface
 {
-    public function
+    public function targetPreprocess(string $target, array $options = []): string;
+    public function getTargetType(string $target): int;
+    public function getFilesWithAbsolutePath(int $targetType, string $target): array;
+    public function getClassesWithNamespaceByFilesWithAbsolutePath(array $filesWithAbsolutePath): array;
+    public function eachClassesWriteDocComment(array $classesWithNamespace, array $ignoreClassesWithNamespace = []): array;
+    public function reportHandleResult(array $scoreBoard): mixed;
 }