+++ src/Console/Command/Generator/PropertyComment/ModeC/Action/ReportResult.php
@@ -2,24 +2,24 @@
 
 declare(strict_types=1);
 
-namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Action;
+namespace Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeC\Action;
 
 use Ghjayce\Shipshape\Action\Action;
-use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeB\Entity\ModeBContext;
+use Ghjayce\Shipshape\Console\Command\Generator\PropertyComment\ModeC\Entity\ModeCContext;
 use Ghjayce\Shipshape\Entity\Context\ClientContext;
 use Ghjayce\Shipshape\Entity\Context\ShipshapeContext;
 
-class ReportScore extends Action
+class ReportResult extends Action
 {
 
     /**
-     * @param ModeBContext $context
+     * @param ModeCContext $context
      * @param ShipshapeContext $shipshapeContext
      * @return mixed
      */
     public function handle(ClientContext $context, ShipshapeContext $shipshapeContext): null
     {
-        $context->getService()->reportHandleResult($context->getScoreBoard());
+        $context->getService()->reportHandleResult($context->getHandleResult());
         return null;
     }
 }